+++ src/components/atoms/CheckInput.tsx
